+++ xroad-catalog-lister/build_rpm.sh
@@ -15,7 +15,7 @@ rm -rf ${ROOT}/RPMS/*
 
 
 rpmbuild \
-    --define "xroad_catalog_version 1.0.0" \
+    --define "xroad_catalog_version 1.0.1" \
     --define "rel $RELEASE" \
     --define "snapshot .$SNAPSHOT" \
     --define "_topdir $ROOT" \